Develop

+++ src/app/profile/page.tsx
@@ -5,21 +5,23 @@ import ProfileButtons from "@/components/buttons/ProfileButtons";
 import FullWrapper from "@/components/page/FullWrapper";
 import PageHeader from "@/components/page/PageHeader";
 import useAuth from "@/components/providers/AuthProvider";
-import CommentProfileTable from "@/components/tables/comment/ProfileTable";
-import RoomProfileTable from "@/components/tables/room/ProfileTable";
-import RoommateProfileTable from "@/components/tables/roommate/ProfileTable";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+import RoomProfilePage from "@/components/tables/room/ProfilePage";
+import RoommateProfileTable from "@/components/tables/roommate/ProfilePage";
 import { Button } from "@/components/ui/button";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { updatePhoto } from "@/firebase/firebaseAuthFunctions";
 import { deleteFile } from "@/firebase/firebaseDBFunctions";
 import { toastMessage } from "@/lib/constants";
 import { sendEmailVerification } from "firebase/auth";
-import { BadgeCheckIcon, BadgeXIcon, HomeIcon, MessageSquareMoreIcon, UsersIcon } from "lucide-react";
+import { BadgeCheckIcon, BadgeXIcon } from "lucide-react";
 import Image from "next/image";
 import { toast } from "sonner";
 import noUserImage from "../../../public/user.png";
+import { useState } from "react";
 
 export default function Profile() {
+  const [tabType, setTabType] = useState("rentals");
   const { currentUser } = useAuth();
 
   const handleSendVerification = async () => {
@@ -33,7 +35,7 @@ export default function Profile() {
 
   const handleRemoveProfilePicture = async () => {
     try {
-      await deleteFile(currentUser?.uid!);
+      if (!currentUser?.photoURL?.includes("google")) await deleteFile(currentUser?.uid!);
       await updatePhoto("");
       await updateProfilePicture(currentUser?.uid!, null);
       toast.success("Profile picture removed successfully");
@@ -51,13 +53,13 @@ export default function Profile() {
         />
         <div className="flex flex-col gap-3 md:flex-row md:items-center lg:gap-5">
           <div className="flex flex-1 gap-3 lg:gap-5">
-            <div className="group relative h-fit w-fit overflow-clip rounded-full">
+            <div className="group relative h-fit w-fit overflow-clip rounded-full border">
               <Image
                 src={currentUser.photoURL ? currentUser.photoURL : noUserImage}
                 alt={currentUser.displayName!}
                 width={50}
                 height={50}
-                className="h-[60px] w-[60px] object-cover md:h-[80px] md:w-[80px] xl:h-[100px] xl:w-[100px]"
+                className="h-[60px] w-[60px] object-cover md:h-[80px] md:w-[80px] xl:h-[90px] xl:w-[90px]"
                 priority
                 placeholder="blur"
                 blurDataURL={currentUser.photoURL ? currentUser.photoURL : ""}
@@ -66,7 +68,7 @@ export default function Profile() {
                 <Button
                   variant="ghost"
                   size="sm"
-                  className="absolute top-0 hidden h-7 w-full border bg-primary-foreground/80 p-2 text-xs group-hover:block"
+                  className="absolute top-0 hidden w-full border bg-primary-foreground/80 pt-1 text-xs group-hover:block md:h-7"
                   onClick={handleRemoveProfilePicture}
                 >
                   Remove
@@ -93,65 +95,47 @@ export default function Profile() {
           </div>
           <ProfileButtons currentUser={currentUser} />
         </div>
-        <Tabs defaultValue="savedAds" className="mt-3 w-full md:mt-0 md:space-y-3">
-          <TabsList>
-            <TabsTrigger value="savedAds">Favourites</TabsTrigger>
-            <TabsTrigger value="postedAds">Posts</TabsTrigger>
-            {/* <TabsTrigger value="comments">Comments</TabsTrigger> */}
-            <TabsTrigger value="reportedAds">Reports</TabsTrigger>
-            {/* <TabsTrigger value="reportedComments">Reported Comments</TabsTrigger> */}
-          </TabsList>
-
-          <TabsContent value="savedAds" className="mt-0 flex flex-col gap-2">
-            <div className="mt-2 flex items-center">
-              <HomeIcon className="mr-1 w-4" />
-              Favourite Room Ads
+        <Tabs defaultValue="savedAds" className="mt-3 w-full md:mt-0 md:space-y-5">
+          <div className="flex justify-between">
+            <TabsList>
+              <TabsTrigger value="savedAds">Favourites</TabsTrigger>
+              <TabsTrigger value="postedAds">Posts</TabsTrigger>
+              <TabsTrigger value="reportedAds">Reports</TabsTrigger>
+            </TabsList>
+            <div className="flex items-center gap-2">
+              <div>Select Ad type</div>
+              <Select value={tabType} onValueChange={setTabType}>
+                <SelectTrigger className="w-[180px]">
+                  <SelectValue placeholder="Select ad type" />
+                </SelectTrigger>
+                <SelectContent>
+                  <SelectItem value="rentals">Rentals</SelectItem>
+                  <SelectItem value="roommates">Roommates</SelectItem>
+                </SelectContent>
+              </Select>
             </div>
-            <RoomProfileTable currentUser={currentUser} tab="savedAds" />
-            <div className="mt-2 flex items-center">
-              <UsersIcon className="mr-1 w-4" />
-              Favourite Roommate Ads
-            </div>
-            <RoommateProfileTable currentUser={currentUser} tab="savedAds" />
+          </div>
+          <TabsContent value="savedAds">
+            {tabType === "rentals" ? (
+              <RoomProfilePage currentUser={currentUser} tab="savedAds" />
+            ) : (
+              <RoommateProfileTable currentUser={currentUser} tab="savedAds" />
+            )}
           </TabsContent>
-          <TabsContent value="postedAds" className="mt-0 flex flex-col gap-2">
-            <div className="mt-2 flex items-center">
-              <HomeIcon className="mr-1 w-4" />
-              Posted Room Ads
-            </div>
-            <RoomProfileTable currentUser={currentUser} tab="postedAds" />
-            <div className="mt-2 flex items-center">
-              <UsersIcon className="mr-1 w-4" />
-              Posted Roommate Ads
-            </div>
-            <RoommateProfileTable currentUser={currentUser} tab="postedAds" />
+          <TabsContent value="postedAds">
+            {tabType === "rentals" ? (
+              <RoomProfilePage currentUser={currentUser} tab="postedAds" />
+            ) : (
+              <RoommateProfileTable currentUser={currentUser} tab="postedAds" />
+            )}
           </TabsContent>
-          {/* <TabsContent value="comments" className="mt-0 flex flex-col gap-2">
-            <div className="mt-2 flex items-center">
-              <MessageSquareMoreIcon className="mr-1 w-4" />
-              Posted Comments
-            </div>
-            <CommentProfileTable currentUser={currentUser} tab="comments" />
-          </TabsContent> */}
-          <TabsContent value="reportedAds" className="mt-0 flex flex-col gap-2">
-            <div className="mt-2 flex items-center">
-              <HomeIcon className="mr-1 w-4" />
-              Reported Room Ads
-            </div>
-            <RoomProfileTable currentUser={currentUser} tab="reportedAds" />
-            <div className="mt-2 flex items-center">
-              <UsersIcon className="mr-1 w-4" />
-              Reported Roommate Ads
-            </div>
-            <RoommateProfileTable currentUser={currentUser} tab="reportedAds" />
+          <TabsContent value="reportedAds">
+            {tabType === "rentals" ? (
+              <RoomProfilePage currentUser={currentUser} tab="reportedAds" />
+            ) : (
+              <RoommateProfileTable currentUser={currentUser} tab="reportedAds" />
+            )}
           </TabsContent>
-          {/* <TabsContent value="reportedComments" className="mt-0 flex flex-col gap-2">
-            <div className="mt-2 flex items-center">
-              <MessageSquareMoreIcon className="mr-1 w-4" />
-              Reported Comments
-            </div>
-            <CommentProfileTable currentUser={currentUser} tab="reportedComments" />
-          </TabsContent> */}
         </Tabs>
       </FullWrapper>
     );
